add backup options.txt - move options helper functions to filehandling
